apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
pydantic-settings==2.7.0
python-json-logger==2.0.7
asyncpg==0.30.0
httpx==0.28.1
PyJWT[crypto]==2.10.1
eval-type-backport==0.2.2

pytest==8.3.4
pytest-asyncio==0.23.8
